Clean up bindigns and sdk
