--- taxel-bindings/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "taxel-bindings"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-build = "build.rs"
-
-[lib]
-path = "src/lib.rs"
-name = "taxel_bindings"
-
-[dependencies]
-taxel-util = { version = "0.1.0", path = "../taxel-util" }
-anyhow = { workspace = true }
-
-[build-dependencies]
-bindgen = "0.60.1"
