+++ taxel-cli/Cargo.toml
@@ -12,7 +12,7 @@ name = "taxel"
 integration-test = []
 
 [dependencies]
-taxel = { version = "0.1.0", path = "../taxel" }
+eric-sdk = "0.3.1"
 taxel-xml = { version = "0.1.0", path = "../taxel-xml" }
 anyhow = { workspace = true }
 log = { workspace = true }
