+++ taxel-cli/src/cmd/mod.rs
@@ -1,6 +1,7 @@
 mod extract;
 mod generate;
 mod send;
+mod utils;
 mod validate;
 
 pub use extract::{extract, extract_args};
