+++ taxel-cli/src/cmd/send.rs
@@ -1,18 +1,21 @@
 //! Validate xml file according to the given taxonomy and send xml file to the
 //! tax authorities.
 
+use super::utils;
 use crate::arg;
 use clap::{Arg, ArgMatches};
-use std::{env::current_dir, path::PathBuf};
-use taxel::{CertificateConfig, Eric, PrintConfig, ProcessingFlag};
+use eric_sdk::Eric;
+use std::{
+    env::current_dir,
+    fs,
+    path::{Path, PathBuf},
+};
 
-pub fn send_args() -> [Arg<'static>; 7] {
+pub fn send_args() -> [Arg<'static>; 5] {
     [
         arg::tax_type(),
         arg::tax_version(),
         arg::xml_file(),
-        arg::certificate_file(),
-        arg::password(),
         arg::print(),
         arg::log_dir(),
     ]
@@ -22,40 +25,25 @@ pub fn send(matches: &ArgMatches) -> Result<(), anyhow::Error> {
     let xml_file = arg::get_one(matches, arg::XML_FILE)?;
     let tax_type = arg::get_one(matches, arg::TAX_TYPE)?;
     let tax_version = arg::get_one(matches, arg::TAX_VERSION)?;
-    let certificate_file = arg::get_one(matches, arg::CERTIFICATE_FILE)?;
-    let password = arg::get_one(matches, arg::PASSWORD)?;
     let log_dir = arg::get_maybe_one(matches, arg::LOG_DIR);
-    let type_version = format!("{}_{}", tax_type, tax_version);
-    let processing_flag: ProcessingFlag;
     let log_path = match log_dir {
         Some(log_dir) => PathBuf::from(log_dir),
         None => current_dir()?,
     };
+    let print_config = if matches.contains_id(arg::PRINT) {
+        let pdf_name = arg::get_one(matches, arg::PRINT)?;
+        Some(pdf_name)
+    } else {
+        None
+    };
+    let xml_path = Path::new(xml_file);
+    let xml = fs::read_to_string(xml_path)?;
 
     let eric = Eric::new(&log_path)?;
 
-    let certificate_config = CertificateConfig::new(certificate_file, password)?;
+    let response = eric.send(xml, tax_type, tax_version, print_config)?;
 
-    let print_config = if matches.contains_id(arg::PRINT) {
-        processing_flag = ProcessingFlag::SendAndPrint;
-        let pdf_name = arg::get_one(matches, arg::PRINT)?;
-        Some(PrintConfig::new(pdf_name, &processing_flag)?)
-    } else {
-        processing_flag = ProcessingFlag::Send;
-        None
-    };
-
-    let xml = taxel::read(xml_file)?;
-
-    let response = eric.send(
-        xml,
-        type_version,
-        processing_flag,
-        certificate_config,
-        print_config,
-    )?;
-
-    eric.log(&log_path, &response)?;
+    utils::log_response(&log_path, &response)?;
 
     Ok(())
 }
