--- taxel/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "taxel"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-
-[lib]
-path = "src/lib.rs"
-name = "taxel"
-
-[features]
-integration-test = []
-
-[dependencies]
-taxel-bindings = { version = "0.1.0", path = "../taxel-bindings" }
-taxel-util = { version = "0.1.0", path = "../taxel-util" }
-anyhow = { workspace = true }
-
-[dev-dependencies]
-roxmltree = { workspace = true }
